drystn_opt: First attempt at adding live-patching that disables EAE (doesn't work currently).
On startup, the program will check for the presence of an EAE unit by multiplying 2x2 and checking if the result in MQ is 4.
If this fails, a set of patches are applied that replace the use of EAE instructions with equivalents that don't require EAE (mostly jumps to separate routines that perform the relevant calculation then jump back to the originating routine after the EAE instruction(s))
After the check, regardless of the outcome, a flag is written to core memory that will inhibit the check from being executed on future runs of the benchmark (until it is fully reloaded again or the flag is cleared manually).