drystn_opt: Rework the EAE patching mechanism.
Before I was prioritizing the EAE implementation of relevant operations, where the EAE version of the code would be inline within the relevant function and the non-EAE implementation would essentially be bodged on top of the EAE implementation.
Usually this entailed replacing one of the instructions within the EAE implementation with a jump to some side routine that did the relevant operation without EAE then jumped back into the core routine after the rest of the EAE code.
This was messy and slightly biased the EAE implementation, since the non-EAE implementations would need to perform two irrelevant jumps that the EAE version did not need to perform.

With this commit, both the EAE and non-EAE implementation of relevant operations are located in side routines, and the main routines include stubs (HLTs) that are replaced with jumps to the correct implementation on startup by the EAE patcher routine, and both impls return to the main routine after the jump.
Now both implementation perform those two extra jumps; neither implementation is biased in this way.
I also feel this is a cleaner way of doing the patching, as both implementation in their entirety always exist, and neither is bodged on top of the other.